+++ src/main/java/kitchenpos/domain/dto/MenuResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.dto;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 
 import java.math.BigDecimal;
 import java.util.List;
