+++ src/main/java/kitchenpos/domain/dto/OrderRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.dto;
 
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.OrderStatus;
 
 import java.util.List;
