+++ src/main/java/kitchenpos/domain/dto/OrderResponse.java
@@ -1,7 +1,7 @@
 package kitchenpos.domain.dto;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
 
 import java.time.LocalDateTime;
 import java.util.List;
