+++ src/main/java/kitchenpos/domain/dto/OrderTableResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.dto;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 
 public class OrderTableResponse {
