+++ src/main/java/kitchenpos/domain/menu/MenuProducts.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import javax.persistence.Embeddable;
 import javax.persistence.FetchType;
