+++ src/main/java/kitchenpos/domain/order/OrderTables.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.TableGroup;
 import org.springframework.util.CollectionUtils;
 
 import javax.persistence.Embeddable;
