+++ src/main/java/kitchenpos/domain/product/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
