+++ src/test/java/kitchenpos/domain/menu/MenuTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.domain.menu;
 
-import kitchenpos.domain.Menu;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
